+++ ehdrs/code.h
@@ -16,7 +16,7 @@ extern Procedure initcodes();
 
 #define RANGE 128 /* ASCII characters are in {0 .. RANGE-1} */
 
-#define Code(c) code_array[c]
+#define Code(c) code_array[(unsigned int)c]
 #define Invcode(code) invcode_array[code]
 
 #ifdef __cplusplus
